职工模块完成

+++ src/main/java/Action/Department/DeptDeleteAction.java
@@ -3,6 +3,7 @@ package Action.Department;
 import Config.MainPanel;
 import Config.MinorFrame;
 import Config.Neo_Button;
+import Config.RemaindLabel;
 import lan.service.DeptService;
 import lan.service.SupermarketService;
 
@@ -30,10 +31,23 @@ public class DeptDeleteAction implements ActionListener{
             @Override
             public void actionPerformed(ActionEvent e) {
                 try{
-                    System.out.println("start dept deleting");
-                    deptService.deleteById(deleteID.getText());
-                    System.out.println("close dept deleting");
-                    menuFrame.dispose();
+                    if(deptService.selectById(deleteID.getText())!= null){
+                        System.out.println("start dept deleting");
+                        deptService.deleteById(deleteID.getText());
+                        System.out.println("close dept deleting");
+                        menuFrame.dispose();
+                    }
+                    else{
+                        menuFrame.dispose();
+                        menuFrame = new MinorFrame("查询结果显示");
+                        JPanel menuPanel = new MainPanel();
+                        menuPanel.setSize(300,200);
+                        RemaindLabel errorMessage = new RemaindLabel("该部门不存在");
+                        menuPanel.add(errorMessage);
+                        menuFrame.setContentPane(menuPanel);
+                        System.out.println("error searching");
+                    }
+
                 }catch (Exception E){}
 
             }
